refactor(app): simplify rules search extraction
